Properly handle adding nullable to $ref's
Combining a $ref with other properties technically is not allowed. The ref should be included in an `allOf` first.

When doing this with UserInteractionModel (which is an array), it completely broke the swift5 code gen. Opted to make that nullable in its definition, because it was nullable everywhere as a safety anyway.